Move export enrichment to async runner
